src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via starlette
boto3==1.25.4
    # via -r requirements.in
boto3-stubs[dynamodb]==1.25.4
    # via -r requirements.in
botocore==1.28.4
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.28.4
    # via boto3-stubs
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via uvicorn
fastapi==0.85.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mypy-boto3-dynamodb==1.25.0
    # via boto3-stubs
pydantic==1.10.2
    # via
    #   -r requirements.in
    #   fastapi
pyhumps==3.8.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
requests==2.28.1
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
types-awscrt==0.15.1
    # via botocore-stubs
types-s3transfer==0.6.0.post4
    # via boto3-stubs
typing-extensions==4.4.0
    # via
    #   boto3-stubs
    #   mypy-boto3-dynamodb
    #   pydantic
    #   starlette
urllib3==1.26.12
    # via
    #   botocore
    #   requests
uvicorn==0.19.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
